feat(appService): moved deactivate endpoint to appChange controller
Ref: CPLP-2644
