+++ tests/marketplace/Apps.Service.Tests/BusinessLogic/AppBusinessLogicTests.cs
@@ -457,29 +457,6 @@ public class AppBusinessLogicTests
 
     #endregion
 
-    #region  DeactivateOfferbyAppId
-
-    [Fact]
-    public async Task DeactivateOfferStatusbyAppIdAsync_CallsExpected()
-    {
-        // Arrange
-        var appId = _fixture.Create<Guid>();
-        var settings = new AppsSettings
-        {
-            ServiceManagerRoles = _fixture.Create<Dictionary<string, IEnumerable<string>>>(),
-            BasePortalAddress = "test"
-        };
-        var sut = new AppsBusinessLogic(null!, null!, _offerService, null!, Options.Create(settings), _mailingService);
-        
-        // Act
-        await sut.DeactivateOfferByAppIdAsync(appId, IamUserId).ConfigureAwait(false);
-
-        // Assert
-        A.CallTo(() => _offerService.DeactivateOfferIdAsync(appId, IamUserId, OfferTypeId.APP)).MustHaveHappenedOnceExactly();
-    }
-
-    #endregion
-
     #region GetAppDocumentContentAsync
 
     [Fact]
